# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Orchids Project Command Center API",
    "description": "Code Style Policeman (CSP) - Team project command center for monitoring development workflow, communication, and code quality heuristics.",
    "version": "1.0.0",
    "contact": {
      "name": "CSP Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "minLength": 1, "maxLength": 100 },
                  "email": { "type": "string", "format": "email", "maxLength": 255 },
                  "password": { "type": "string", "minLength": 8, "maxLength": 128, "description": "Must contain uppercase, lowercase, and a number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" },
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "409": { "description": "Email already in use" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/api/webhooks/github": {
      "post": {
        "summary": "GitHub webhook receiver (push, pull_request, issues, deployment_status)",
        "tags": ["Webhooks"],
        "parameters": [
          { "name": "workspace_id", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Webhook processed" },
          "401": { "description": "Invalid signature" }
        }
      }
    },
...
    "/api/webhooks/whatsapp": {
      "post": {
        "summary": "WhatsApp message relay receiver",
        "tags": ["Webhooks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["workspace_id", "from", "body"],
                "properties": {
                  "workspace_id": { "type": "string", "format": "uuid" },
                  "from": { "type": "string" },
                  "body": { "type": "string" },
                  "timestamp": { "type": "string" },
                  "group_name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Message processed" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string" },
          "name": { "type": "string" },
          "avatar_url": { "type": "string", "nullable": true },
          "github_username": { "type": "string", "nullable": true }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" }
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limited",
        "headers": {
          "Retry-After": { "schema": { "type": "integer" }, "description": "Seconds until rate limit resets" }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
}